Update QuickSort.h
int to double

+++ QuickSort.h
@@ -3,7 +3,7 @@ class QuickSort
 {
 public:
 	void swap(int* a, int* b);
-	void quickSort(int array[], int low, int high);
-	int partition(int array[], int low, int high);
+	void quickSort(double array[], double low, double high);
+	double partition(double array[], double low, double high);
 };
